+++ suite-common/suite-sync/src/labeling/updateAddressLabel.ts
@@ -1,7 +1,7 @@
 import { UpdateAddressLabel } from '@suite-common/suite-sync-types';
 import { ok } from '@trezor/type-utils';
 
-import { EnsureStorageDep } from '../storage/ensureStorage';
+import { EnsureStorageDep } from '../storage/createEnsureStorage';
 
 export type UpdateAddressLabelDeps = EnsureStorageDep;
