+++ suite-common/suite-sync/src/labeling/updateWalletLabel.ts
@@ -2,7 +2,7 @@ import { UpdateWalletLabel } from '@suite-common/suite-sync-types';
 import { parseDeviceStaticSessionId } from '@suite-common/wallet-utils';
 import { ok } from '@trezor/type-utils';
 
-import { EnsureStorageDep } from '../storage/ensureStorage';
+import { EnsureStorageDep } from '../storage/createEnsureStorage';
 
 export type UpdateWalletLabelDeps = EnsureStorageDep;
